mise a la norme de printf (separations des fonctions dans plusieurs fichiers)
